blogdown {blogdown}	R Documentation

The blogdown package

Description

The comprehensive documentation of this package is the book blogdown: Creating Websites with R Markdown (https://bookdown.org/yihui/blogdown/). You are expected to read at least the first chapter. If you are really busy or do not care about an introduction to blogdown (e.g., you are very familiar with creating websites), set your working directory to an empty directory, and run blogdown::new_site() to get started right away.

Examples

if (interactive()) blogdown::new_site()

[Package blogdown version 0.17 Index]