build_dir {blogdown}	R Documentation

Build all Rmd files under a directory

Description

List all Rmd files recursively under a directory, and compile them using rmarkdown::render().

Usage

build_dir(dir = ".", force = FALSE, ignore = "[.]Rproj$")

Arguments

dir
A directory path.

force
Whether to force building all Rmd files. By default, an Rmd file is built only if it is newer than its output file(s).

ignore
A regular expression to match output filenames that should be ignored when testing if the modification time of the Rmd source file is newer than its output files.

[Package blogdown version 0.17 Index]