dep_path {blogdown}	R Documentation

A helper function to return a dependency path name

Description

In most cases, blogdown can process images and HTML widgets automatically generated from code chunks (they will be moved to the static/ folder by default), but it may fail to recognize dependency files generated to other paths. This function returns a path that you can use for your output files, so that blogdown knows that they should be be processed, too. It is designed to be used in a knitr code chunk.

Usage

dep_path(default = knitr::opts_chunk$get("fig.path"))

Arguments

default
Return this default value when this function is called outside of a knitr code chunk.

Value

A character string of the default value (outside knitr), or a path consisting of the knitr figure path appended by the current chunk label.

[Package blogdown version 0.17 Index]