install_hugo {blogdown}	R Documentation

Install Hugo

Description

Download the appropriate Hugo executable for your platform from Github and try to copy it to a system directory so blogdown can run the hugo command to build a site. update_hugo() is a wrapper of install_hugo(force = TRUE).

Usage

install_hugo(version = "latest", use_brew = Sys.which("brew") != "", force = FALSE)

update_hugo()

Arguments

version
The Hugo version number, e.g., 0.26; the special value latest means the latest version (fetched from Github releases).

use_brew
Whether to use Homebrew (https://brew.sh) on macOS to install Hugo (recommended if you have already installed Homebrew). Note Homebrew will be automatically installed if it has not been installed when use_brew = TRUE.

force
Whether to install Hugo even if it has already been installed. This may be useful when upgrading Hugo (if you use Homebrew, run the command brew update && brew upgrade instead).

Details

This function tries to install Hugo to Sys.getenv('APPDATA') on Windows, ‘~/Library/Application Support’ on macOS, and ‘~/bin/’ on other platforms (such as Linux). If these directories are not writable, the package directory ‘Hugo’ of blogdown will be used. If it still fails, you have to install Hugo by yourself and make sure it can be found via the environment variable PATH.

This is just a helper function and may fail to choose the correct Hugo executable for your operating system, especially if you are not on Windows or Mac or a major Linux distribution. When in doubt, read the Hugo documentation and install it by yourself: https://gohugo.io.

[Package blogdown version 0.5 Index]